my camera moved from conference into component
